Refactor UI into ControlPanel
Part of #495
